chore: update test cases from problem specifications for prime-factors (#895)
- Updated all the test cases including new ones added
- Updated examples to also use `long long` instead of `int`
[no important files changed]

+++ exercises/practice/prime-factors/.meta/example.h
@@ -3,10 +3,9 @@
 
 #include <vector>
 
-namespace prime_factors
-{
+namespace prime_factors {
 
-std::vector<int> of(int n);
+std::vector<long long> of(long long n);
 
 }
